% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviar.R
\name{enviar}
\alias{enviar}
\title{Enviar email ABJ}
\usage{
enviar(arq, from = "jtrecenti@abj.org.br", dests = c("reunioes@abj.org.br"))
}
\arguments{
\item{arq}{Arquivo Rmd.}

\item{from}{quem enviou o e-mail? Default jtrecenti at abj.org.br.}

\item{dests}{vetor de emails dos destinatários. Default reunioes at abj.org.br.}
}
\description{
Envia um e-mail com base em arquivo Rmd de reuniões.
}

