% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_meet}
\alias{new_meet}
\title{Abrir template}
\usage{
new_meet(nome, dir = "reunioes", data = format(Sys.Date(), "\%Y\%m\%d"))
}
\arguments{
\item{nome}{Titulo da reuniao.}

\item{dir}{Pasta onde os arquivos serão salvos.}

\item{data}{data em formato YYYYMMDD. Default data atual.}
}
\description{
Abrir template de reunioes.
}

